move template survey files to Survey subdirectory
