+++ awx/ui_next/src/screens/Template/Survey/SurveyList.jsx
