+++ awx/ui_next/src/screens/Template/Survey/SurveyListItem.jsx
