+++ awx/ui_next/src/screens/Template/Survey/index.js
@@ -1,0 +1,3 @@
+export { default as SurveyList } from './SurveyList';
+export { default as SurveyQuestionAdd } from './SurveyQuestionAdd';
+export { default as SurveyQuestionEdit } from './SurveyQuestionEdit';
